+++ src/basic.js
@@ -9,6 +9,10 @@
  *  . none,
  *
  *
+ * Private Static Methods:
+ *  . _setTestMode                returns internal objects for testing purpose,
+ *
+ *
  * Public Static Methods:
  *  . noConflict                  returns a reference to this ES6lib object,
  *  . getString                   returns a string,
@@ -16,7 +20,7 @@
  *
  *
  *
- * @namespace    ES6lib
+ * @namespace    -
  * @dependencies none
  * @exports      -
  * @author       -
@@ -24,7 +28,7 @@
  * @version      -
  * ************************************************************************ */
 /* global Tree, root */
-/* - */
+/* eslint-disable no-underscore-dangle */
 
 'use strict';
 
@@ -48,9 +52,32 @@
   // -- Local Variables
 
 
-  // -- Public Static Methods ------------------------------------------------
+  // -- Main -----------------------------------------------------------------
 
   ES6lib = {
+
+    // Useful to retrieve the library name and version when it is
+    // embedded in another library as an object:
+    library: { name: '{{lib:name}}', version: '{{lib:version}}' },
+
+
+    // -- Private Static Methods ---------------------------------------------
+
+    /**
+     * Returns the internal objects for testing purpose.
+     *
+     * @method ()
+     * @private
+     * @param {}            -,
+     * @returns {Object}    returns a list of internal objects,
+     * @since 0.0.0
+     */
+    _setTestMode() {
+      return [];
+    },
+
+
+    // -- Public Static Methods ----------------------------------------------
 
     /**
      * Returns a reference to this ES6lib object.
@@ -104,4 +131,4 @@
 
   // END OF IIFE
 }());
-/* - */
+/* eslint-enable no-underscore-dangle */