+++ src/functional-shared.js
@@ -14,6 +14,10 @@
  *  . ES6lib                      creates and returns the ES6lib object,
  *
  *
+ * Private Static Methods:
+ *  . _setTestMode                returns internal objects for testing purpose,
+ *
+ *
  * Public Static Methods:
  *  . noConflict                  returns a reference to this ES6lib object,
  *
@@ -24,7 +28,7 @@
  *
  *
  *
- * @namespace    ES6lib
+ * @namespace    -
  * @dependencies none
  * @exports      -
  * @author       -
@@ -32,7 +36,7 @@
  * @version      -
  * ************************************************************************ */
 /* global Tree, root, extend */
-/* eslint-disable one-var, semi-style */
+/* eslint-disable one-var, semi-style, no-underscore-dangle */
 
 'use strict';
 
@@ -73,6 +77,10 @@
    */
   ES6lib = function(name) {
     const obj = {};
+    obj.library = {
+      name: '{{lib:name}}',
+      version: '{{lib:version}}',
+    };
     obj.name = name;
     extend(obj, methods);
     return obj;
@@ -80,6 +88,22 @@
 
   // Attaches a constant to ES6lib that provides the version of the lib.
   ES6lib.VERSION = '{{lib:version}}';
+
+
+  // -- Private Static Methods -----------------------------------------------
+
+  /**
+   * Returns the internal objects for testing purpose.
+   *
+   * @method ()
+   * @private
+   * @param {}              -,
+   * @returns {Object}      returns a list of internal objects,
+   * @since 0.0.0
+   */
+  ES6lib._setTestMode = function() {
+    return [];
+  };
 
 
   // -- Public Static Methods ------------------------------------------------
@@ -138,4 +162,4 @@
 
   // END OF IIFE
 }());
-/* eslint-enable one-var, semi-style */
+/* eslint-enable one-var, semi-style, no-underscore-dangle */