+++ src/pseudoclassical.js
@@ -14,6 +14,10 @@
  *  . ES6lib                      creates and returns the ES6lib object,
  *
  *
+ * Private Static Methods:
+ *  . _setTestMode                returns internal objects for testing purpose,
+ *
+ *
  * Public Static Methods:
  *  . noConflict                  returns a reference to this ES6lib object,
  *
@@ -24,7 +28,7 @@
  *
  *
  *
- * @namespace    ES6lib
+ * @namespace    -
  * @dependencies none
  * @exports      -
  * @author       -
@@ -32,7 +36,7 @@
  * @version      -
  * ************************************************************************ */
 /* global Tree, root */
-/* eslint-disable one-var, semi-style */
+/* eslint-disable one-var, semi-style, no-underscore-dangle */
 
 'use strict';
 
@@ -73,11 +77,31 @@
     if (!(this instanceof ES6lib)) {
       throw new Error('ES6lib needs to be called with the new keyword!');
     }
+    this.library = {
+      name: '{{lib:name}}',
+      version: '{{lib:version}}',
+    };
     this.name = name;
   };
 
   // Attaches a constant to ES6lib that provides the version of the lib.
   ES6lib.VERSION = '{{lib:version}}';
+
+
+  // -- Private Static Methods -----------------------------------------------
+
+  /**
+   * Returns the internal objects for testing purpose.
+   *
+   * @method ()
+   * @private
+   * @param {}              -,
+   * @returns {Object}      returns a list of internal objects,
+   * @since 0.0.0
+   */
+  ES6lib._setTestMode = function() {
+    return [];
+  };
 
 
   // -- Public Static Methods ------------------------------------------------
@@ -136,4 +160,4 @@
 
   // END OF IIFE
 }());
-/* eslint-enable one-var, semi-style */
+/* eslint-enable one-var, semi-style, no-underscore-dangle */